requirements.txt
asgiref==3.2.10
click==7.1.2
Django==3.1.2
djangorestframework==3.12.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
mysqlclient==2.0.1
protobuf==3.13.0
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
Werkzeug==1.0.1
